+++ addon/mixins/observable.js
@@ -1,41 +1,51 @@
 import Ember from 'ember';
+import { Observable } from 'rxjs/Observable';
 import { Subject } from 'rxjs/Subject';
 import { BehaviorSubject } from 'rxjs/BehaviorSubject';
 import { merge } from 'rxjs/observable/merge';
+import { Subscription } from 'rxjs/Subscription';
 import 'rxjs/add/operator/map';
 import 'rxjs/add/operator/skip';
+import 'rxjs/add/operator/multicast';
 
 function action(actionName) {
-  let existing = this.actions[actionName] || function() {};
-  let s = this._actionSubjects[actionName];
+  let s = this._actionObservables[actionName];
   if (!s) {
-     s = (this._actionSubjects[actionName] = new Subject());
-     this.actions[actionName] = function() {
-       try {
-         let result = existing.apply(this, arguments);
-         s.next(arguments[0]);
-         return result;
-       } catch(e) {
-         s.error(e);
-       }
-     };
+    s = Observable.create(o => {
+      let existing = this.actions[actionName];
+      this.actions[actionName] = function(x) {
+        try {
+          let r = (existing || function() {}).apply(this, arguments);
+          o.next(x);
+          return r;
+        } catch (e) {
+          o.error(e);
+        }
+      };
+      return new Subscription(() => {
+        this.actions[actionName] = existing;
+      });
+    }).multicast(new Subject()).refCount();
+
+    this._actionObservables[actionName] = s;
   }
-  return s.asObservable();
+  return s;
 }
 
 function property(propertyName) {
-  let s = this._propSubjects[propertyName];
+  let s = this._propertyObservables[propertyName];
   if (!s) {
-     s = (this._propSubjects[propertyName] = new BehaviorSubject(this.get(propertyName)));
-     this.addObserver(propertyName, () => {
-       try {
-         s.next(this.get(propertyName));
-       } catch (e) {
-         s.error(e);
-       }
-     });
+    s = Observable.create(o => {
+      let next = () => o.next(this.get(propertyName));
+      this.addObserver(propertyName, next);
+      return new Subscription(() => {
+        this.removeObserver(propertyName, next);
+      });
+    }).multicast(new BehaviorSubject(this.get(propertyName))).refCount();
+
+    this._propertyObservables[propertyName] = s;
   }
-  return s.asObservable();
+  return s;
 }
 
 function properties(...props) {
@@ -55,8 +65,8 @@ export default Ember.Mixin.create({
   init() {
     this._super(...arguments);
 
-    this._actionSubjects = {};
-    this._propSubjects = {};
+    this._actionObservables = {};
+    this._propertyObservables = {};
 
     this.observable = {
       action: (actionName) => action.call(this, actionName),