+++ tests/unit/mixins/observable-test.js
@@ -23,6 +23,22 @@ test('this.observable.property', function(assert) {
   subject.set('x', '3');
 });
 
+test('this.observable.property unsubscription', function(assert) {
+
+  let ObserveActionObject = Ember.Object.extend(ObservableMixin);
+  let subject = ObserveActionObject.create({ x: '1' });
+  assert.ok(subject);
+
+  assert.equal(subject.observersForKey('x').length, 0);
+
+  let sub = subject.observable.property('x').subscribe(() => {});
+  assert.equal(subject.observersForKey('x').length, 1);
+
+  sub.unsubscribe();
+  assert.equal(subject.observersForKey('x').length, 0);
+
+});
+
 test('this.observable.properties', function(assert) {
   let done = assert.async();
 
@@ -42,6 +58,25 @@ test('this.observable.properties', function(assert) {
   subject.set('x', '2');
 });
 
+test('this.observable.properties unsubscription', function(assert) {
+
+  let ObserveActionObject = Ember.Object.extend(ObservableMixin);
+  let subject = ObserveActionObject.create({ x: '1' });
+  assert.ok(subject);
+
+  assert.equal(subject.observersForKey('x').length, 0);
+  assert.equal(subject.observersForKey('y').length, 0);
+
+  let sub = subject.observable.properties('x','y').subscribe(() => {});
+  assert.equal(subject.observersForKey('x').length, 1);
+  assert.equal(subject.observersForKey('y').length, 1);
+
+  sub.unsubscribe();
+  assert.equal(subject.observersForKey('x').length, 0);
+  assert.equal(subject.observersForKey('x').length, 0);
+
+});
+
 test('this.observable.action', function(assert) {
   let done = assert.async();
 
@@ -56,4 +91,24 @@ test('this.observable.action', function(assert) {
   });
 
   subject.send('test', { x: 1 });
+});
+
+test('this.observable.action unsubscription', function(assert) {
+  let ObserveActionRoute = Ember.Route.extend(ObservableMixin, {
+    actions: {
+      test() { }
+    }
+  });
+
+  let subject = ObserveActionRoute.create();
+  assert.ok(subject);
+
+  let originalAction = subject.actions.test;
+  assert.equal(typeof originalAction, 'function');
+
+  let sub = subject.observable.action('test').subscribe(() => {});
+  assert.notEqual(subject.actions.test, originalAction);
+
+  sub.unsubscribe();
+  assert.equal(subject.actions.test, originalAction);
 });